Minus bags

+++ cmd/Berkat.go
@@ -13,7 +13,6 @@ import (
 	"time"
 	telegram "web-scraper/tg-bot"
 
-	"github.com/go-co-op/gocron"
 	"github.com/gocolly/colly"
 	"github.com/joho/godotenv"
 )
@@ -37,11 +36,12 @@ func init() {
 
 func main() {
 	BotToken = os.Getenv("BOT_TOKEN")
-	s := gocron.NewScheduler(time.UTC)
-	// добавляем одну задачу на каждые 10 минут
-	s.Cron("*/10 * * * *").Do(task)
-	// запускаем планировщик с блокировкой текущего потока
-	s.StartBlocking()
+	// s := gocron.NewScheduler(time.UTC)
+	// // добавляем одну задачу на каждые 10 минут
+	// s.Cron("*/10 * * * *").Do(task)
+	// // запускаем планировщик с блокировкой текущего потока
+	// s.StartBlocking()
+	task()
 }
 
 // func testTask(){
@@ -51,10 +51,16 @@ func task() {
 	fmt.Println("Начало")
 	keys := GetKeys()
 	c := telegram.New(BotToken)
-	for _, value := range keys {
-		err := c.SendMessage("Доступен новый ключ \xF0\x9F\x94\x91 \n\n\xF0\x9F\x93\x8D Локация: "+value.Country+"\n\xF0\x9F\x8C\x90 Трафик: \xE2\x99\xBE \n\n\xF0\x9F\x94\x92 Тип: "+value.Type+"  ```\n"+value.Key+"```", int64(-1002343650923))
-		if err != nil {
-			fmt.Println(err)
+	if len(keys) != 0 {
+		for i := len(keys) - 1; i >= 0; i-- {
+			if keys[i].Key != "" {
+				err := c.SendMessage("Доступен новый ключ \xF0\x9F\x94\x91 \n\n\xF0\x9F\x93\x8D Локация: "+keys[i].Country+"\n\xF0\x9F\x8C\x90 Трафик: \xE2\x99\xBE \n\n\xF0\x9F\x94\x92 Тип: "+keys[i].Type+"  ```\n"+keys[i].Key+"```", int64(-1002343650923))
+				if err!=nil  {
+					fmt.Println("Cлишком много ключей")
+					time.Sleep(2 * time.Minute)
+				}
+				time.Sleep(1 * time.Second)
+			}
 		}
 	}
 	fmt.Println("Конец")
@@ -66,7 +72,7 @@ func GetKeys() []Object {
 	c := colly.NewCollector()
 	// set a valid User-Agent header
 	c.UserAgent = "Mozilla/5.0 (compatible; MSIE 9.0; Windows; U; Windows NT 6.1; Win64; x64 Trident/5.0)"
-	keys := make([]Object, 1)
+	var keys []Object
 
 	file, err := os.Open("cmd/fer.txt")
 	if err != nil {
@@ -88,10 +94,7 @@ func GetKeys() []Object {
 	file.Close()
 
 	var Continue bool = true
-	keys[0].Id = lastID + 1
-
-	for i := 1; lastID != keys[len(keys)-1].Id; i++ {
-
+	for i := 1; len(keys) == 0 || lastID < keys[len(keys)-1].Id; i++ {
 		c.OnHTML(".col", func(e *colly.HTMLElement) {
 			var key Object
 			var err error
@@ -111,7 +114,8 @@ func GetKeys() []Object {
 			if key.Online == "" {
 				key.Online = "Offline"
 			}
-			if key.Id == lastID {
+			if key.Id <= lastID && Continue {
+
 				err := os.Remove("cmd/fer.txt")
 				if err != nil {
 					fmt.Println("Ошибка при удалении файла:", err)
@@ -124,14 +128,17 @@ func GetKeys() []Object {
 					os.Exit(1)
 				}
 				defer file.Close()
-
-				file.WriteString(strconv.Itoa(keys[1].Id))
-				Continue = false
+				if len(keys) == 0 {
+					file.WriteString(strconv.Itoa(key.Id))
+					Continue = false
+				} else {
+					file.WriteString(strconv.Itoa(keys[0].Id))
+					Continue = false
+				}
 			}
 			if Continue {
 				keys = append(keys, key)
 			}
-
 		})
 		num := strconv.Itoa(i)
 		c.Visit("https://outlinekeys.com/?page=" + num)
@@ -139,12 +146,11 @@ func GetKeys() []Object {
 			break
 		}
 	}
+
 	var SendKeys []Object
-	for key, value := range keys {
+
+	for _, value := range keys {
 		if value.Country == "Russia" || value.Type == "Vless" || value.Online == "Offline" {
-			continue
-		}
-		if key == 0 {
 			continue
 		}
 		SendKeys = append(SendKeys, value)
@@ -152,14 +158,16 @@ func GetKeys() []Object {
 	for num, value := range SendKeys {
 		href := "/key/" + strconv.Itoa(value.Id) + "/"
 		vpnkey := GetVpnKey(href)
-		SendKeys[num].Key = vpnkey[0]
-		fmt.Println(SendKeys[num])
+		if len(vpnkey) > 0 {
+			SendKeys[num].Key = vpnkey[0]
+			fmt.Println(SendKeys[num])
+		}
 	}
 	return SendKeys
 }
 
 func GetVpnKey(url string) (surs []string) {
-	time.Sleep(2 * time.Second)
+	time.Sleep(1 * time.Second)
 	g := colly.NewCollector()
 	g.UserAgent = "Mozilla/5.0 (compatible; MSIE 9.0; Windows; U; Windows NT 6.1; Win64; x64 Trident/5.0)"
 	g.OnHTML(".form-control", func(e *colly.HTMLElement) {
